Move the api related views to the api/views.py.
